Remove body overflow hidden and replace with service to prevent scrolling
This commit is in response to https://github.com/likeastore/ngDialog/issues/156.

It removes css for clipping page content with overflow:hidden on `html` and `body` as this causes issues with endless scrolling and is generally bad UX.

As a replacement, I've added a service that is able to consume all scroll events and block them from propagating or at worst reverting their effects. This is all done with only core Angular and Javascript.

I have tested this on a large enterprise application with very good results.